requirements.txt
streamlit>=1.32.0
openai>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
python-docx>=1.1.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
fpdf2>=2.7.0
openpyxl>=3.1.2
python-dateutil>=2.8.0
